A Recurrent Neural Network Music Generation Tutorial
This model provides baselines for the application of language modeling
to melody generation. This code also serves as a working example for
implementing a language model in TensorFlow.

There are two parts to this change:
1) magenta/models/scripts and magenta/models/lib contains melody
extraction code.
2) magenta/models/basic_rnn/ contains the model.

The melody extraction code takes NoteSequence protos (created from MIDI
files) and saves melodies it finds in each instrument track. The
melodies are monophonic, meaning one note at a time.

The model is an LSTM implemented in TensorFlow which is trained to
predict the next note in a sequence. There is also a generator that
samples melodies from the distribution the model learns.