Misc README link edit (#60)
When reading the readme for basic_rnn I had to stop and see where convert_sequences_to_melodies.py is, but if we instead link to the folder and not directly to it's readme, then those files will be visible right at the top. Provides good context for the readme, mirrors how all github repos are layed out.