Simplify to_tensors logic in music_vae data pipeline. Part 1.
Define tf_to_tensor() and to_tensors() out of BaseConverter as module-level fns that are invoked by get_dataset(). BaseConverter.to_tensor() points to the new convert_to_tensors() and is not overridden by any subclass to simplify the flow (for now). This removes much of the interweaving logic in tf_to_tensors and makes the code that is shared by all converters explicit. For instance, currently OneHotMelodyConverter.tf_to_tensor() has the following flow:

OneHotMelodyConverter.tf_to_tensor()
-> BaseConverter.tf_to_tensor()
-> BaseNoteSequenceConverter.to_tensor()
-> BaseConverter.to_tensor()
-> OneHotMelodyConverter._to_tensor()
-> LegacyEventListOneHotConverter._to_tensor()

which is simplified to the following with this CL:
tf_to_tensor()
-> BaseConverter.to_tensor()
-> convert_to_tensors()
-> OneHotMelodyConverter._to_tensor()
-> LegacyEventListOneHotConverter._to_tensor()

Ultimately, we should be able to delete these base classes completely, make BaseConverter an abstract class, and eliminate interweaving logic.

Upcoming TODOs:
+ Decouple other functions from base classes (sampling, combining, etc.)
+ Decouple to_notesequences() similar to to_tensors()
+ Consolidate __init__ for subclasses
+ Remove base legacy and base converters
+ Cleanup

PiperOrigin-RevId: 303844329