Consolidate encoders in music_vae DataConverters.
This change consolidates to_tensors and _to_tensors to expose a single to_tensors() method. Additionally, all base implementations are moved to util fns so that the encoding logic for each final DataConverter is contained entirely within that class (no interweaving, easier to read, maintain and extend). This makes the intermediate base classes abstract, and most of them will be removed subsequently.

PiperOrigin-RevId: 305112151