[numpy] Fix users of NumPy APIs that are removed in NumPy 2.0.
This change migrates users of APIs removed in NumPy 2.0 to their recommended replacements (https://numpy.org/doc/stable/numpy_2_0_migration_guide.html).

This change replaces uses of `np.math`, which is a deprecated alias for the builtin `math` module, with the builtin `math` module. `np.math` is removed in NumPy 2.0.

PiperOrigin-RevId: 658227917